Fixed incorrect signature of WEL_DISPATCHER_DELEGATE due to changes in WEL   from INTEGER to POINTER so that it compiled under 64 bits.
git-svn-id: https://svn.eiffel.com/eiffelstudio/trunk@48944 8089f293-4706-0410-a29e-feb5c42a2edf

+++ Src/bench/Eiffel/eiffel/com_il_generation/Core/run-time/runtime.cs
@@ -16,7 +16,7 @@ using EiffelSoftware.Runtime.Types;
 
 namespace EiffelSoftware.Runtime {
 
-public delegate int WEL_DISPATCHER_DELEGATE (IntPtr hwnd, int msg, int wparam, int lparam);
+public delegate int WEL_DISPATCHER_DELEGATE (IntPtr hwnd, int msg, IntPtr wparam, IntPtr lparam);
 public delegate void WEL_ENUM_FONT_DELEGATE (IntPtr lpelf, IntPtr lpntm, int font_type, IntPtr lparam);
 public delegate void WEL_ENUM_WINDOW_DELEGATE (IntPtr hwnd);
 public delegate void EV_PIXMAP_IMP_DELEGATE (int error_code, int data_type, int pixmap_width,
